Enhanced Model Selection and Configuration Management
This update introduces optional use of the "--config" argument, with default values employed if it is not provided. A new YAML configuration file has been implemented to manage default parameters for various language learning models. Additionally, the command center's code has been revised to dynamically adjust the action engine based on the model selected in the UI, thereby extending support to a wider range of models including OpenAI GPT-3.5, HuggingFace Mixtral-8x7B, OpenAI LiteLLM-GPT-3.5-Turbo, and OpenAI GPT-4 (AZURE).
